% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_create.R
\name{pack_create}
\alias{pack_create}
\title{Wrapper for passing DESCRIPTION fields and creating a package directory}
\usage{
pack_create(pkg_name = "test_package", rstudio = TRUE, open = FALSE,
  ...)
}
\arguments{
\item{pkg_name}{package name as string, field for DESCRIPTION}

\item{rstudio}{Creates an RStudio project. Default is TRUE}

\item{open}{Open the project in RStudio on creation. Default is FALSE}

\item{...}{Pass any other variables to usethis::create_package()}

\item{path}{directory path as string, field for DESCRIPTION}

\item{first}{First author name, field for DESCRIPTION}

\item{last}{Last author name, field for DESCRIPTION}

\item{email}{Author email for CRAN, field for DESCRIPTION}

\item{role}{, field for DESCRIPTION. Default 'c("aut", "cre")'}

\item{lic}{LICENCE to use. e.g. GPL-3, MIT, Apache2.0, CCBY4, etc. Default is "GPL-3",
field for DESCRIPTION}

\item{lang}{Field for language, field for DESCRIPTION. Default is "en-GB"}
}
\value{
Creates the directory structure in the path provided.
}
\description{
pack_create() is a wrapper on usethis::create_pacakge() Passes
field arguments for DESCRIPTION and creates the directory skeleton for an R package.
}
\note{
Choose a good package name! See e.g. <\url{http://r-pkgs.had.co.nz/package.html#naming}>.
For authorship see for example:
https://journal.r-project.org/archive/2012-1/RJournal_2012-1_Hornik~et~al.pdf
This function will generally create the following files and folders
in the specified path:
.Rbuildignore # not if rstudio = FALSE
.Rproj.user # not if rstudio = FALSE
.gitignore # empty
DESCRIPTION
NAMESPACE
R # empty
pkg_name.Rproj # not if rstudio = FALSE
man # this directory now gets created when running roxygen
}
\examples{

\dontrun{
pack_create(pkg_name = 'test_package',
            path = '/tmp/',
            first = "Super",
            last = "Duper",
            email = "super@duper.com",
            role = 'c("aut", "cre")',
            lic = "GPL-3",
            lang = "en-GB"
            )


}

}
\seealso{
\code{\link{functioname}},
\code{\link[usethis]{create_package}}.
}
\author{
Antonio J. Berlanga-Taylor <\url{https://github.com/AntonioJBT/Rpacker}>
}
