% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpac_all_setup.R
\name{rpac_all_setup}
\alias{rpac_all_setup}
\title{Setup templates and directories for an R package}
\usage{
rpac_all_setup(pkg_name = NULL, path = ".", first = NULL,
  last = NULL, email = NULL, pkgs = NULL, github_user = NULL)
}
\arguments{
\item{pkg_name}{package name as string}

\item{path}{Pass a path as string. Default is current working directory ('.').}

\item{first}{First author name, string}

\item{last}{Last author name, string}

\item{email}{Author email for CRAN, string}

\item{pkgs}{Packages to add to DESCRIPTION, pass as string.}

\item{github_user}{Pass as string, will be inserted into
"https://github.com/github_user/"}
}
\description{
rpac_all_setup() creates various directories and files to quickly
setup an R package. Includes skeleton structure for code, tests, documentation
plus various templates. Various defaults assumed, including templates for
GitHub, Travis-CI and Codecov.
}
\examples{

\dontrun{
# Travel to parent directory for setup:
setwd('/somewhere/sensible/like/github_repos/')

# Quick setup for R package directory structure and several templates:
rpac_all_setup(pkg_name = 'packageTest',
               path = '.',
               first = "Super",
               last = "Duper",
               email = "super@duper.com",
               pkgs = 'dummy_holder',
               github_user = 'SuperDuper'
               )

}

}
\seealso{
\code{\link{rpac_create}},
\code{\link{rpac_rbuildignore}},
\code{\link{rpac_dependencies}},
\code{\link{rpac_gitignore}},
\code{\link{rpac_travis}},
\code{\link{rpac_cran_comments}},
\code{\link{rpac_readme}},
\code{\link[usethis]{use_package_doc}},
\code{\link[usethis]{use_vignette}},
\code{\link[usethis]{use_testthat}},
\code{\link[usethis]{use_coverage}}.
}
\author{
Antonio J Berlanga-Taylor <\url{https://github.com/AntonioJBT/Rpacker}>
}
