% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpac_codecov.R
\name{rpac_codecov}
\alias{rpac_codecov}
\title{Create a template for code coverage yaml file}
\usage{
rpac_codecov(path = ".", append = TRUE, threshold = "1\%")
}
\arguments{
\item{path}{Pass a path as string. Default is current working directory ('.').}

\item{append}{Append text to an existing file. Default is TRUE.}

\item{threshold}{codecov threshold to pass. Default is 1%.}
}
\description{
rpac_codecov() creates a yaml template for use with covr and
codecov
}
\examples{

\dontrun{
rpac_codecov()

}

}
\seealso{
<\url{https://codecov.io/}>
}
\author{
Antonio J Berlanga-Taylor <\url{https://github.com/AntonioJBT/Rpacker}>
}
