% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpac_cran_comments.R
\name{rpac_cran_comments}
\alias{rpac_cran_comments}
\title{Create a text file template for CRAN submission}
\usage{
rpac_cran_comments(path = ".", append = TRUE)
}
\arguments{
\item{path}{Pass a path as string. Default is current working directory ('.').}

\item{append}{Append text to an existing file. Default is TRUE.}
}
\description{
rpac_cran_comments() creates a template to log errors, warnings
and notes to make it easier for CRAN submission. Wraps usethis::use_cran_comments()
and adds some text to it.
}
\examples{

\dontrun{
rpac_cran_comments()

}

}
\seealso{
\code{\link[usethis]{use_cran_comments}}.
}
\author{
Antonio J Berlanga-Taylor <\url{https://github.com/AntonioJBT/Rpacker}>
}
