% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpac_dependencies.R
\name{rpac_dependencies}
\alias{rpac_dependencies}
\title{Add packages to DESCRIPTION}
\usage{
rpac_dependencies(pkgs = NULL, type = "Suggests", ...)
}
\arguments{
\item{pkgs}{Packages to add to DESCRIPTION, pass as strings}

\item{type}{Type of dependency from use_package(). Default is 'Suggests'}

\item{...}{Pass any other arguments to use_package()}
}
\value{

}
\description{
rpac_dependencies() adds packages to load which are dependencies.
Light wrapper for usethis::use_package()
}
\note{
Prefer 'Suggests' but use 'Imports' if many functions rely on it and you dont'
}
\examples{

\dontrun{
pkgs <- c('data.table',
          'ggplot2',
          'cowplot',
          'dplyr'
          )
rpac_dependencies(pkgs = pkgs)

}

}
\seealso{
<\url{http://r-pkgs.had.co.nz/description.html#dependencies}>,
\code{\link[usethis]{use_package}}.
}
\author{
Antonio J Berlanga-Taylor <\url{https://github.com/AntonioJBT/Rpacker}>
}
