% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpac_function.R
\name{rpac_function}
\alias{rpac_function}
\title{Create a template for a function in roxygen2 style}
\usage{
rpac_function(function_name = NULL, pkg_name = NULL, path = "./R",
  author = NULL, github_user = NULL, level = "short", open = TRUE)
}
\arguments{
\item{function_name}{Function name, used to create file name. Appends '.R'.}

\item{pkg_name}{package name as string.}

\item{path}{Pass a path as string. Default is to place files for functions in
the directory './R'.}

\item{author}{Author name as string.}

\item{github_user}{Pass as string, will be inserted into
https://github.com/github_user/}

\item{level}{'short' provides placeholders; 'long' adds explanations.}

\item{open}{Open the project in RStudio on creation. Default is TRUE.}
}
\value{

}
\description{
rpac_function() creates a file with a template to create a function.
Follows roxygen2 conventions and intended to be used with devtools.
}
\note{
level 'short' is not so short.
}
\examples{

\dontrun{
rpac_function(function_name = 'rpac_test_long2',
              pkg_name = 'Rpacker',
              path = 'R',
              author = 'Antonio',
              github_user = 'antonio',
              level = 'long',
              open = TRUE # tries to open file
              )

}

}
\seealso{
<\url{http://r-pkgs.had.co.nz/r.html}>
}
\author{
Antonio J Berlanga-Taylor <\url{https://github.com/AntonioJBT/Rpacker}>
}
