% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpac_gitignore.R
\name{rpac_gitignore}
\alias{rpac_gitignore}
\title{Create a .gitignore file template}
\usage{
rpac_gitignore(path = ".", append = TRUE)
}
\arguments{
\item{path}{Pass a path as string. Default is current working directory ('.').}

\item{append}{Append text to an existing file. Default is TRUE.}
}
\description{
rpac_gitignore() Creates a preset .gitignore file.
}
\examples{

\dontrun{
rpac_gitignore()


}

}
\author{
Antonio J Berlanga-Taylor <\url{https://github.com/AntonioJBT/Rpacker}>
}
