% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpac_rbuildignore.R
\name{rpac_rbuildignore}
\alias{rpac_rbuildignore}
\title{Creates a .Rbuildignore template file}
\usage{
rpac_rbuildignore(path = ".", append = TRUE)
}
\arguments{
\item{path}{Pass a path as string. Default is current working directory ('.').}

\item{append}{Append text to an existing file. Default is TRUE}
}
\value{
Creates a .Rbuildignore file in the working directory
}
\description{
rpac_rbuildignore() creates a .Rbuildignore file
}
\note{
Append is false, this will overwrite a file with the same name.
}
\examples{

\dontrun{
setwd('/my_github/my_packages/test_package')
rpac_rbuildignore()

}

}
\author{
Antonio J Berlanga-Taylor <\url{https://github.com/AntonioJBT/Rpacker}>
}
