% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpac_readme.R
\name{rpac_readme}
\alias{rpac_readme}
\title{Generates a README template for a GitHub repository}
\usage{
rpac_readme(github_user = NULL, pkg_name = NULL, path = ".",
  append = TRUE)
}
\arguments{
\item{github_user}{Pass as string, will be inserted into
"https://github.com/github_user/"}

\item{pkg_name}{package name as string}

\item{path}{Pass a path as string. Default is current working directory ('.').}

\item{append}{Append text to an existing file. Default is TRUE.}
}
\value{
Creates a README.md file in the working directory
}
\description{
rpac_readme() creates a README, includes badges travisCI and
codecov
}
\examples{

\dontrun{
rpac_readme(github_user = 'me',
            pkg_name = 'superPackage',
            path = '.'
            )
}

}
\author{
Antonio J Berlanga-Taylor <\url{https://github.com/AntonioJBT/Rpacker}>
}
