% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpac_test.R
\name{rpac_test_template}
\alias{rpac_test_template}
\alias{rpac_add_test}
\title{Create a template for a test in testthat style and add boilerplate code}
\usage{
rpac_test_template(test_name = NULL, pkg_name = NULL,
  path = "tests/testthat")

rpac_add_test(test_name = NULL, function_name = NULL,
  path = "tests/testthat/", append = TRUE, open = TRUE)
}
\arguments{
\item{test_name}{Name for test, inserted into 'test-%s.R' to create a file
name in rpac_test_template(). rpac_add_test() expects the same string to find
the file and append to it.}

\item{pkg_name}{package name as string.}

\item{path}{Pass a path as string. Default is to place files for tests in
the directory './tests/testthat/'.}

\item{function_name}{Name of function that test will be used for.}

\item{append}{Append text to an existing file. Default is TRUE.}

\item{open}{Open the project in RStudio on creation. Default is TRUE.}
}
\description{
rpac_test_template() creates a file with boilerplate code
for running tests with testthat. rpac_add_test() appends a blurb for tests for
a given function.
}
\note{
Run this for each file needed to create a test-testname.R
file with a blurb. Files must be within tests/testthat/ or inst/tests
and start with 'test' and end in '.R' for testthat to pick them up.
There are many types of tests, see testthat::expect_.
}
\examples{

\dontrun{
rpac_test_template(test_name = 'rpack_test_template',
                   pkg_name = 'rpack',
                   path = 'tests/testthat/'
                   )

rpac_add_test(test_name = 'rpack_test_template',
              function_name = 'function_to_test',
              path = 'tests/testthat',
              append = TRUE,
              open = TRUE
              )
}

}
\seealso{
<\url{http://r-pkgs.had.co.nz/tests.html}>,
<\url{http://testthat.r-lib.org/articles/custom-expectation.html}>,
\code{\link[usethis]{use_test}}.
}
\author{
Antonio J. Berlanga-Taylor <\url{https://github.com/AntonioJBT/Rpacker}>
}
