% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpac_travis.R
\name{rpac_travis}
\alias{rpac_travis}
\title{Create a Travis-CI yaml template}
\usage{
rpac_travis(path = ".", append = TRUE)
}
\arguments{
\item{path}{Pass a path as string. Default is current working directory ('.').}

\item{append}{Append text to an existing file. Default is TRUE.}
}
\description{
rpac_travis() creates a template for Linux and Mac OS tests with
suggestions for settings for covr and vdiffr testing.
}
\note{
Adds blurbs for travis-ci and codecov shields. Packages need to be
added manually to .travis.yml and DESCRIPTION later on.
Also see the template from usethis::use_travis()
}
\examples{

\dontrun{
rpac_travis()
# Copy packages manually as functions and dependencies get added:
pkgs <- c('data.table',
          'ggplot2',
          'cowplot',
          'dplyr'
          )
for (i in pkgs) {
  i <- paste0('- ', i)
  cat(i, '\\n')
  }
}

}
\seealso{
\code{\link[usethis]{use_travis}}.
}
\author{
Antonio J Berlanga-Taylor <\url{https://github.com/AntonioJBT/Rpacker}>
}
