% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpacker-package.R
\docType{package}
\name{Rpacker-package}
\alias{Rpacker}
\alias{Rpacker-package}
\title{Rpacker: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Antonio J. Berlanga-Taylor \email{antoniojberlanga@gmail.com}

}
\keyword{internal}
